+++ CHANGES.md
@@ -1,3 +1,9 @@
+2015-04-15
+
+* Refactored create-list-of-cases.rb
+* create-list-of-cases.rb now warns if scenario files don't exist
+
+
 2015-04-06
 
 * Fixed errors in generated emissions constraint files